# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Network Navigator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Connection {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateConnectionBody {
  name: string;
  /** @nullable */
  notes?: string | null;
  /**
   * Optional follow-up date to schedule immediately
   * @nullable
   */
  followUpDate?: string | null;
}

export interface UpdateConnectionBody {
  name?: string;
  /** @nullable */
  notes?: string | null;
}

export interface FollowUp {
  id: number;
  connectionId: number;
  userId: string;
  scheduledDate: string;
  originalDueDate: string;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
  connection: Connection;
}

export interface CreateFollowUpBody {
  connectionId: number;
  scheduledDate: string;
}

export interface UpdateFollowUpBody {
  scheduledDate?: string;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardSummary {
  todayCount: number;
  upcomingCount: number;
  overdueCount: number;
  weekFollowUps: FollowUp[];
  overdueFollowUps: FollowUp[];
  upcomingFollowUps: FollowUp[];
  /** Follow-ups due today */
  todayFollowUps: FollowUp[];
  /** Long-term follow-ups (30+ day span) due within the next 5 days */
  reminderFollowUps: FollowUp[];
}

export interface ProgressStats {
  totalScheduled: number;
  totalCompleted: number;
  onTimeCount: number;
  /** On-time follow-through rate as a percentage (0-100). onTimeCount / totalScheduled * 100 */
  followThroughRate: number;
  /** Total completion rate as a percentage (0-100). totalCompleted / totalScheduled * 100 */
  completionRate: number;
  completedFollowUps: FollowUp[];
}

export type ListFollowUpsParams = {
  connectionId?: number;
  status?: ListFollowUpsStatus;
};

export type ListFollowUpsStatus =
  (typeof ListFollowUpsStatus)[keyof typeof ListFollowUpsStatus];

export const ListFollowUpsStatus = {
  pending: "pending",
  completed: "completed",
  overdue: "overdue",
} as const;